+++ brokerapi/brokers/gcp_service_broker.go
@@ -494,10 +494,9 @@ func getStaticPlans() (map[string][]models.ServicePlan, error) {
 	for _, p := range plans {
 		serviceId := p["service_id"].(string)
 		planName := p["name"].(string)
-
-		id, err := db_service.GetOrCreatePlanId(planName, serviceId)
-		if err != nil {
-			return map[string][]models.ServicePlan{}, err
+		planId, planIdOk := p["id"]
+		if !planIdOk {
+			return map[string][]models.ServicePlan{}, fmt.Errorf("Error: plan ids are required. Plan %s needs an id.", p["name"].(string))
 		}
 
 		servicePropertyBytes, err := json.Marshal(p["service_properties"])
@@ -519,7 +518,7 @@ func getStaticPlans() (map[string][]models.ServicePlan, error) {
 				DisplayName: p["display_name"].(string),
 				Bullets:     []string{p["description"].(string), "For pricing information see https://cloud.google.com/pricing/#details"},
 			},
-			ID:                id,
+			ID:                planId.(string),
 			ServiceProperties: servicePropertyMap,
 		}
 
@@ -536,7 +535,7 @@ func getStaticPlans() (map[string][]models.ServicePlan, error) {
 				ServiceId: serviceId,
 				Name:      p["name"].(string),
 				Features:  string(servicePropertyBytes),
-				ID:        id,
+				ID:        planId.(string),
 			}
 			db_service.DbConnection.Create(&planDetails)
 		}
@@ -565,10 +564,9 @@ func getDynamicPlans(envVarName string, translatePlanFunc func(details map[strin
 		// save custom plans to database and construct mapping
 		for planName, planDetails := range dynamicPlans {
 			serviceId = planDetails["service"]
-
-			id, err := db_service.GetOrCreatePlanId(planName, planDetails["service"])
-			if err != nil {
-				return []models.ServicePlan{}, "", err
+			planId, planIdOk := planDetails["id"]
+			if !planIdOk {
+				return []models.ServicePlan{}, "", fmt.Errorf("Error: plan ids are required. Plan %s needs an id.", planDetails["name"])
 			}
 
 			// get service-specific plan service_properties from plan interface
@@ -596,7 +594,7 @@ func getDynamicPlans(envVarName string, translatePlanFunc func(details map[strin
 					ServiceId: planDetails["service"],
 					Name:      planDetails["name"],
 					Features:  string(servicePropertiesStr),
-					ID:        id,
+					ID:        planId,
 				}
 				db_service.DbConnection.Create(&existingPlan)
 			}